% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_forecast_to_fisch_format}
\alias{convert_forecast_to_fisch_format}
\title{convert_forecast_to_fisch_format}
\usage{
convert_forecast_to_fisch_format(forecast)
}
\arguments{
\item{forecast}{vector of forecasted inflows of length equal to length(inflow)}
}
\description{
Prepare model inputs by snapping to desired disretization level.
}
