% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\name{DP}
\alias{DP}
\title{DP}
\usage{
DP(
  n_periods,
  inflow_forecast,
  S_upper_limit,
  S_lower_limit,
  R_disc_x,
  S_states,
  State_mat,
  S_end_state,
  price_forecast,
  Rev_to_go,
  Bellman,
  Policy,
  use_storage_target
)
}
\arguments{
\item{n_periods}{number of sub-daily periods}

\item{S_upper_limit}{upper limit of storage}

\item{S_lower_limit}{lower limit of storage}

\item{R_disc_x}{release discretization}

\item{State_mat}{storage state matric}

\item{Rev_to_go}{placeholder matrix for revenue to go}

\item{Bellman}{placeholder matrix for Bellman's function}

\item{Policy}{placeholder matrix for release policy}

\item{S_States}{storage stages}
}
\description{
Weekly release and storage operations guided by ISTARF wrapping around DP sub-daily scheduling.
}
