% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisch-package.R
\docType{package}
\name{fisch}
\alias{fisch-package}
\alias{fisch}
\title{fisch: Forecast Informed SCheduler for Hydropower}
\description{
fisch is a hydropower scheduler that optimizes hourly releases within a week to maximize hydropower revenue. Dynamic programing is used to identify the optimal release schedule given hourly electricity prices, hourly inflow, and an end-of-week storage target that parameterizes non-powered objectives and constraints.
}
