% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_firomax.R
\name{schedule_release}
\alias{schedule_release}
\title{schedule_release}
\usage{
schedule_release(
  mode = "single_fixed",
  inflow = TEST_inflow,
  inflow_forecast = TEST_inflow_forecast,
  price_forecast = TEST_price_forecast,
  MWh_per_MCM = TEST_MWh_per_MCM,
  initial_storage = TEST_initial_storage,
  reservoir_capacity = TEST_reservoir_capacity,
  target_end_of_week_storage = TEST_target_end_of_week_storage,
  max_release = TEST_max_release,
  min_release = TEST_min_release,
  discretization_step = 0.1,
  use_storage_target = TRUE
)
}
\arguments{
\item{mode}{"single_fixed" => single forecast used at start of week; release is based on that forecast | "single_adaptive" => single forecast used at start of week, but release responds to policy during simulation (i.e., updates with storage) | "rolling_adaptive" => policy is updated daily with new forecast information; simulation is rolling horizon.}

\item{inflow}{actual inflow time series.}

\item{inflow_forecast}{inflow forecast with daily updating and same resolution as inflow. If NULL, perfect forecast is applied.}

\item{initial_storage}{storage level at start of week.}

\item{reservoir_capacity}{res}

\item{target_end_of_week_storage}{storage level targeted for end of week}

\item{max_release}{maximum}

\item{min_release}{minimum}

\item{discretization_step}{default 0.1 (allowable levels 0.0001, 0.001, 0.01, 0.1, 1, 10)}

\item{use_storage_target}{T/F}

\item{price_foreacst}{price forecast with same temporal resolution as inflow.}

\item{release_value}{power value of turbined water in MWh per MCM.}
}
\description{
Optimizes then simulates one week of sub-daily release.
}
