% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{snap_to_disc}
\alias{snap_to_disc}
\title{snap_to_disc}
\usage{
snap_to_disc(x, discretization_step)
}
\arguments{
\item{x}{data to be disretized (either scalar or vector)}

\item{discretization_step}{default 0.2. Other water volume variables (inflow, inflow_forecast, reservoir_capacity, etc.) must all be discretized at this level.}
}
\description{
Prepare model inputs by snapping to desired disretization level.
}
