% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisch-package.R
\docType{package}
\name{fisch-package}
\alias{fisch}
\alias{fisch-package}
\title{fisch: Forecast Informed Reservoir Operations for MAXimum hydroelectric power sales}
\description{
fisch performs Dynamic Programming to optimize within week reservoir releases for maximum hydropower revenue. Weekly release targets from the `ISTARF-CONUS` dataset may be used to soft-constrain operations. Sub-daily release decisions are optimized in full-week or rolling-horizon mode for given deterministic price and streamflow forecasts.
}
\keyword{internal}
